requirements.txt
accelerate==0.34.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
-e git+https://github.com/keirp/automatic_prompt_engineer.git@eac521c79a78965245ce7745dcc9f6b0792c7ec7#egg=automatic_prompt_engineer
backoff==2.2.1
backports.tarfile==1.2.0
bayesian-optimization==1.5.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bitarray==2.9.2
bitsandbytes==0.41.1
bleach==6.1.0
blessed==1.20.0
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cma==4.0.0
cmake==3.30.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
Comet==3.1.0
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
coverage==7.6.1
cryptography==40.0.2
cycler==0.12.1
Cython==3.0.11
datasets==2.13.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
easse @ git+https://github.com/feralvam/easse.git@6a4352ec299ed03fda8ee45445ca43d9c7673e89
entmax==1.3
et-xmlfile==1.1.0
executing==2.0.1
fairseq==0.12.2
fastapi==0.112.0
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.15.4
fire==0.6.0
FLAML==2.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gpustat==1.1.1
gradio==4.44.0
gradio_client==1.3.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
hydra-core==1.0.7
hyperlink==21.0.0
identify==2.6.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
jaraco.context==5.3.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonargparse==3.13.1
jsonlines==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
kubernetes==30.1.0
Levenshtein==0.25.1
lightning-utilities==0.11.7
lit==18.1.8
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.4.0
mosestokenizer==1.2.1
mpmath==1.3.0
msal==1.20.0
multidict==6.0.5
multiprocess==0.70.14
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nevergrad==0.7.0
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.0.6
onnxruntime==1.18.1
openai==0.28.0
openfile==0.0.7
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
optuna==3.6.1
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.2.240909
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
portalocker==2.10.1
posthog==3.5.0
pre-commit==3.8.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
-e git+https://github.com/dongyuanjushi/autogen.git@b3e4e7045b1c694251fb41987bcf69ae96466d6c#egg=pyautogen
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pytest==7.4.4
pytest-asyncio==0.23.8
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0
pyzmq==26.1.0
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.7.24
replicate==0.31.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rouge==1.0.1
rpds-py==0.20.0
rsa==4.9
ruff==0.6.4
sacrebleu==2.3.1
sacremoses==0.1.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.1.99
shapely==2.0.5
shellingham==1.5.4
simalign==0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
stack-data==0.6.3
stanfordnlp==0.2.0
starlette==0.37.2
sympy==1.13.2
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
tomlkit==0.12.0
toolwrapper==2.1.0
torch==2.1.0
torchaudio==2.1.0
torchmetrics==0.10.3
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@e39b6c1c7cdc890b6849b8c9de545fc9590ba871
triton==2.1.0
-e git+https://github.com/sylinrl/TruthfulQA@fdd8ad1c0d00a478cf8b0bb41a3ad8378c16293b#egg=truthfulqa
tseval @ git+https://github.com/facebookresearch/text-simplification-evaluation.git@dea8863683ea5946fd50184883c9be7a7339e821
Twisted==24.7.0
typer==0.12.3
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
uctools==1.3.0
unbabel-comet==2.2.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
vertexai==1.60.0
virtualenv==20.26.3
watchfiles==0.23.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wolframalpha==5.1.3
wrapt==1.16.0
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.11.1
yattag==1.16.0
zipp==3.20.0
zope.interface==7.0.3
